Update : Changing Most Skill.ID from ushort to int Update : Most Derived Constructor from Skill to handle int "ID" and remove ushort cast. Update : Packet Lib should now be the only place where ID are sent as "ushort" Update : Dictionary to handle disabled skill now reflect this change to "int" ID.
Nothing should change in actual behavior when ID were cast to ushort before even storing them in Object.
Relations between subspell ID and spell ID could get better with this.
If you have spell ID that collide once converted to ushort it could be a trouble for using them in Quickbar or GameEffect Icon.
It Should be no trouble for "Mob Only" Spells that are cast only on mobs (that include charm and pet) or that doesn't spawn an effect icon !

git-svn-id: https://svn.code.sf.net/p/dolserver/code/DOLSharp/trunk@3355 c9877f0d-d2ea-46ae-93cb-121acdf3e971

+++ GameServer/gameutils/Ability.cs
@@ -69,12 +69,12 @@ namespace DOL.GS
 		{
 		}
 
-		public Ability(string keyname, string displayname, string description, ushort id, ushort icon, int level)
+		public Ability(string keyname, string displayname, string description, int id, ushort icon, int level)
 			: this(keyname, displayname, description, id, icon, level, "", 0)
 		{
 		}
 
-		public Ability(string keyname, string displayname, string description, ushort id, ushort icon, int level, string spec, int speclevel)
+		public Ability(string keyname, string displayname, string description, int id, ushort icon, int level, string spec, int speclevel)
 			: base(keyname, displayname, id, icon, level)
 		{
 			m_spec = spec;
